Backend/requirements.txt
# Requires Python 3.12.6 or newer to run the application
python>=3.12.6

# Flask 3.1.0+ provides the web framework for routing, sessions, and templating
flask>=3.1.0

# MySQL Connector/Python 9.2.0+ for connecting to MySQL databases
mysql-connector-python>=9.2.0

# bcrypt 4.3.0+ for secure password hashing and verification
bcrypt>=4.3.0

# pdfkit 1.0.0+ Python wrapper to convert HTML templates into PDF files
pdfkit>=1.0.0

# wkhtmltopdf (0.2+) system package required by pdfkit to render PDFs from HTML
wkhtmltopdf>=0.2

# Werkzeug 3.1.3+ provides WSGI utilities including secure password hashing tools
Werkzeug>=3.1.3


